+++ Tests/Helpers/XRPTransaction+Test.swift
@@ -1,5 +1,8 @@
 @testable import XpringKit
 
 extension XRPTransaction {
-  public static let testTransaction = XRPTransaction(getTransactionResponse: .testGetTransactionResponse)!
+  public static let testTransaction = XRPTransaction(
+    getTransactionResponse: .testGetTransactionResponse,
+    xrplNetwork: XRPLNetwork.test
+    )!
 }