+++ Tests/XRP/DefaultXRPClientTest.swift
@@ -11,7 +11,10 @@ final class DefaultXRPClientTest: XCTestCase {
   // MARK: - Balance
   func testGetBalanceWithSuccess() {
     // GIVEN an XRPClient which will successfully return a balance from a mocked network call.
-    let xrpClient = DefaultXRPClient(networkClient: FakeNetworkClient.successfulFakeNetworkClient)
+    let xrpClient = DefaultXRPClient(
+      networkClient: FakeNetworkClient.successfulFakeNetworkClient,
+      xrplNetwork: XRPLNetwork.test
+    )
 
     // WHEN the balance is requested.
     guard let balance = try? xrpClient.getBalance(for: .testAddress) else {
@@ -29,7 +32,10 @@ final class DefaultXRPClientTest: XCTestCase {
       XCTFail("Failed to decode X-Address.")
       return
     }
-    let xrpClient = DefaultXRPClient(networkClient: FakeNetworkClient.successfulFakeNetworkClient)
+    let xrpClient = DefaultXRPClient(
+      networkClient: FakeNetworkClient.successfulFakeNetworkClient,
+      xrplNetwork: XRPLNetwork.test
+    )
 
     // WHEN the balance is requested THEN an error is thrown.
     XCTAssertThrowsError(
@@ -55,7 +61,7 @@ final class DefaultXRPClientTest: XCTestCase {
       transactionStatusResult: .success(.testGetTransactionResponse),
       transactionHistoryResult: .success(.testTransactionHistoryResponse)
     )
-    let xrpClient = DefaultXRPClient(networkClient: networkClient)
+    let xrpClient = DefaultXRPClient(networkClient: networkClient, xrplNetwork: XRPLNetwork.test)
 
     // WHEN the balance is requested THEN the error is thrown.
     XCTAssertThrowsError(try xrpClient.getBalance(for: .testAddress), "Exception not thrown") { error in
@@ -71,7 +77,10 @@ final class DefaultXRPClientTest: XCTestCase {
   // MARK: - Send
   func testSendWithSuccess() {
     // GIVEN an XRPClient client which will successfully return a balance from a mocked network call.
-    let xrpClient = DefaultXRPClient(networkClient: FakeNetworkClient.successfulFakeNetworkClient)
+    let xrpClient = DefaultXRPClient(
+      networkClient: FakeNetworkClient.successfulFakeNetworkClient,
+      xrplNetwork: XRPLNetwork.test
+    )
 
     // WHEN XRP is sent.
     guard
@@ -95,7 +104,10 @@ final class DefaultXRPClientTest: XCTestCase {
       XCTFail("Failed to decode X - Address.")
       return
     }
-    let xrpClient = DefaultXRPClient(networkClient: FakeNetworkClient.successfulFakeNetworkClient)
+    let xrpClient = DefaultXRPClient(
+      networkClient: FakeNetworkClient.successfulFakeNetworkClient,
+      xrplNetwork: XRPLNetwork.test
+    )
 
     // WHEN XRP is sent to a classic address THEN an error is thrown.
     XCTAssertThrowsError(
@@ -109,7 +121,10 @@ final class DefaultXRPClientTest: XCTestCase {
 
   func testSendWithInvalidAddress() {
     // GIVEN an XRPClient client and an invalid destination address.
-    let xrpClient = DefaultXRPClient(networkClient: FakeNetworkClient.successfulFakeNetworkClient)
+    let xrpClient = DefaultXRPClient(
+      networkClient: FakeNetworkClient.successfulFakeNetworkClient,
+      xrplNetwork: XRPLNetwork.test
+    )
     let destinationAddress = "xrp"
 
     // WHEN XRP is sent to an invalid address THEN an error is thrown.
@@ -131,7 +146,7 @@ final class DefaultXRPClientTest: XCTestCase {
       transactionStatusResult: .success(.testGetTransactionResponse),
       transactionHistoryResult: .success(.testTransactionHistoryResponse)
     )
-    let xrpClient = DefaultXRPClient(networkClient: networkClient)
+    let xrpClient = DefaultXRPClient(networkClient: networkClient, xrplNetwork: XRPLNetwork.test)
 
     // WHEN a send is attempted then an error is thrown.
     XCTAssertThrowsError(
@@ -152,7 +167,7 @@ final class DefaultXRPClientTest: XCTestCase {
       transactionStatusResult: .success(.testGetTransactionResponse),
       transactionHistoryResult: .success(.testTransactionHistoryResponse)
     )
-    let xrpClient = DefaultXRPClient(networkClient: networkClient)
+    let xrpClient = DefaultXRPClient(networkClient: networkClient, xrplNetwork: XRPLNetwork.test)
 
     // WHEN a send is attempted then an error is thrown.
     XCTAssertThrowsError(
@@ -173,7 +188,7 @@ final class DefaultXRPClientTest: XCTestCase {
       transactionStatusResult: .success(.testGetTransactionResponse),
       transactionHistoryResult: .success(.testTransactionHistoryResponse)
     )
-    let xrpClient = DefaultXRPClient(networkClient: networkClient)
+    let xrpClient = DefaultXRPClient(networkClient: networkClient, xrplNetwork: XRPLNetwork.test)
 
     // WHEN a send is attempted then an error is thrown.
     XCTAssertThrowsError(
@@ -202,7 +217,7 @@ final class DefaultXRPClientTest: XCTestCase {
         transactionStatusResult: .success(transactionStatusResponse),
         transactionHistoryResult: .success(.testTransactionHistoryResponse)
       )
-      let xrpClient = DefaultXRPClient(networkClient: networkClient)
+      let xrpClient = DefaultXRPClient(networkClient: networkClient, xrplNetwork: XRPLNetwork.test)
 
       // WHEN the payment status is retrieved.
       let paymentStatus = try? xrpClient.paymentStatus(for: .testTransactionHash)
@@ -225,7 +240,7 @@ final class DefaultXRPClientTest: XCTestCase {
       transactionStatusResult: .success(transactionStatusResponse),
       transactionHistoryResult: .success(.testTransactionHistoryResponse)
     )
-    let xrpClient = DefaultXRPClient(networkClient: networkClient)
+    let xrpClient = DefaultXRPClient(networkClient: networkClient, xrplNetwork: XRPLNetwork.test)
 
     // WHEN the payment status is retrieved.
     let paymentStatus = try? xrpClient.paymentStatus(for: .testTransactionHash)
@@ -249,7 +264,7 @@ final class DefaultXRPClientTest: XCTestCase {
         transactionStatusResult: .success(transactionStatusResponse),
         transactionHistoryResult: .success(.testTransactionHistoryResponse)
       )
-      let xrpClient = DefaultXRPClient(networkClient: networkClient)
+      let xrpClient = DefaultXRPClient(networkClient: networkClient, xrplNetwork: XRPLNetwork.test)
 
       // WHEN the payment status is retrieved.
       let paymentStatus = try? xrpClient.paymentStatus(for: .testTransactionHash)
@@ -272,7 +287,7 @@ final class DefaultXRPClientTest: XCTestCase {
       transactionStatusResult: .success(transactionStatusResponse),
       transactionHistoryResult: .success(.testTransactionHistoryResponse)
     )
-    let xrpClient = DefaultXRPClient(networkClient: networkClient)
+    let xrpClient = DefaultXRPClient(networkClient: networkClient, xrplNetwork: XRPLNetwork.test)
 
     // WHEN the payment status is retrieved.
     let paymentStatus = try? xrpClient.paymentStatus(for: .testTransactionHash)
@@ -290,7 +305,7 @@ final class DefaultXRPClientTest: XCTestCase {
       transactionStatusResult: .failure(XpringKitTestError.mockFailure),
       transactionHistoryResult: .success(.testTransactionHistoryResponse)
     )
-    let xrpClient = DefaultXRPClient(networkClient: networkClient)
+    let xrpClient = DefaultXRPClient(networkClient: networkClient, xrplNetwork: XRPLNetwork.test)
 
     // WHEN the payment status is retrieved THEN an error is thrown.
     XCTAssertThrowsError(try xrpClient.paymentStatus(for: .testTransactionHash))
@@ -308,7 +323,7 @@ final class DefaultXRPClientTest: XCTestCase {
       transactionStatusResult: .success(getTransactionResponse),
       transactionHistoryResult: .success(.testTransactionHistoryResponse)
     )
-    let xrpClient = DefaultXRPClient(networkClient: networkClient)
+    let xrpClient = DefaultXRPClient(networkClient: networkClient, xrplNetwork: XRPLNetwork.test)
 
     // WHEN the payment status is retrieved.
     let paymentStatus = try? xrpClient.paymentStatus(for: .testTransactionHash)
@@ -334,7 +349,7 @@ final class DefaultXRPClientTest: XCTestCase {
       transactionStatusResult: .success(getTransactionResponse),
       transactionHistoryResult: .success(.testTransactionHistoryResponse)
     )
-    let xrpClient = DefaultXRPClient(networkClient: networkClient)
+    let xrpClient = DefaultXRPClient(networkClient: networkClient, xrplNetwork: XRPLNetwork.test)
 
     // WHEN the payment status is retrieved.
     let paymentStatus = try? xrpClient.paymentStatus(for: .testTransactionHash)
@@ -347,7 +362,10 @@ final class DefaultXRPClientTest: XCTestCase {
 
   func testAccountExistsWithSuccess() {
     // GIVEN an XRPClient which will successfully return a balance from a mocked network call.
-    let xrpClient = DefaultXRPClient(networkClient: FakeNetworkClient.successfulFakeNetworkClient)
+    let xrpClient = DefaultXRPClient(
+      networkClient: FakeNetworkClient.successfulFakeNetworkClient,
+      xrplNetwork: XRPLNetwork.test
+    )
 
     // WHEN the existence of the account is checked.
     guard let exists = try? xrpClient.accountExists(for: .testAddress) else {
@@ -365,7 +383,10 @@ final class DefaultXRPClientTest: XCTestCase {
       XCTFail("Failed to decode X-Address.")
       return
     }
-    let xrpClient = DefaultXRPClient(networkClient: FakeNetworkClient.successfulFakeNetworkClient)
+    let xrpClient = DefaultXRPClient(
+      networkClient: FakeNetworkClient.successfulFakeNetworkClient,
+      xrplNetwork: XRPLNetwork.test
+    )
 
     // WHEN the account's existence is checked THEN an error is thrown.
     XCTAssertThrowsError(
@@ -401,7 +422,7 @@ final class DefaultXRPClientTest: XCTestCase {
       transactionStatusResult: .success(.testGetTransactionResponse),
       transactionHistoryResult: .success(.testTransactionHistoryResponse)
     )
-    let xrpClient = DefaultXRPClient(networkClient: networkClient)
+    let xrpClient = DefaultXRPClient(networkClient: networkClient, xrplNetwork: XRPLNetwork.test)
 
     // WHEN the existence of the account is checked
     guard let exists = try? xrpClient.accountExists(for: .testAddress) else {
@@ -433,7 +454,7 @@ final class DefaultXRPClientTest: XCTestCase {
       transactionStatusResult: .success(.testGetTransactionResponse),
       transactionHistoryResult: .success(.testTransactionHistoryResponse)
     )
-    let xrpClient = DefaultXRPClient(networkClient: networkClient)
+    let xrpClient = DefaultXRPClient(networkClient: networkClient, xrplNetwork: XRPLNetwork.test)
 
     // WHEN the account's existence is checked THEN the error is re-thrown.
     XCTAssertThrowsError(
@@ -453,12 +474,15 @@ final class DefaultXRPClientTest: XCTestCase {
 
   func testPaymentHistoryWithSuccess() {
     // GIVEN an XRPClient client which will successfully return a transactionHistory mocked network call.
-    let xrpClient = DefaultXRPClient(networkClient: FakeNetworkClient.successfulFakeNetworkClient)
+    let xrpClient = DefaultXRPClient(
+      networkClient: FakeNetworkClient.successfulFakeNetworkClient,
+      xrplNetwork: XRPLNetwork.test
+    )
     let expectedTransactions = Org_Xrpl_Rpc_V1_GetAccountTransactionHistoryResponse
       .testTransactionHistoryResponse
       .transactions
       .map { transactionResponse in
-        return XRPTransaction(getTransactionResponse: transactionResponse)
+        return XRPTransaction(getTransactionResponse: transactionResponse, xrplNetwork: XRPLNetwork.test)
       }
 
     // WHEN the transactionHistory is requested.
@@ -477,7 +501,10 @@ final class DefaultXRPClientTest: XCTestCase {
       XCTFail("Failed to decode X-Address.")
       return
     }
-    let xrpClient = DefaultXRPClient(networkClient: FakeNetworkClient.successfulFakeNetworkClient)
+    let xrpClient = DefaultXRPClient(
+      networkClient: FakeNetworkClient.successfulFakeNetworkClient,
+      xrplNetwork: XRPLNetwork.test
+    )
 
     // WHEN the payment history is requested THEN an error is thrown.
     XCTAssertThrowsError(
@@ -503,7 +530,7 @@ final class DefaultXRPClientTest: XCTestCase {
       transactionStatusResult: .success(.testGetTransactionResponse),
       transactionHistoryResult: .failure(XpringKitTestError.mockFailure)
     )
-    let xrpClient = DefaultXRPClient(networkClient: networkClient)
+    let xrpClient = DefaultXRPClient(networkClient: networkClient, xrplNetwork: XRPLNetwork.test)
 
     // WHEN the payment history is requested THEN an error is thrown.
     XCTAssertThrowsError(try xrpClient.paymentHistory(for: .testAddress), "Exception not thrown") { error in
@@ -522,7 +549,7 @@ final class DefaultXRPClientTest: XCTestCase {
     // Generate expected transactions from the default response, which only contains payments.
     var transactionHistory = Org_Xrpl_Rpc_V1_GetAccountTransactionHistoryResponse.testTransactionHistoryResponse
     let expectedTransactions = transactionHistory.transactions.map { transactionResponse in
-      return XRPTransaction(getTransactionResponse: transactionResponse)
+      return XRPTransaction(getTransactionResponse: transactionResponse, xrplNetwork: XRPLNetwork.test)
     }
 
     // Append a non-payment transaction. This is not one of the expected outputs because it is not a payment.
@@ -539,7 +566,7 @@ final class DefaultXRPClientTest: XCTestCase {
       transactionStatusResult: .success(.testGetTransactionResponse),
       transactionHistoryResult: .success(transactionHistory)
     )
-    let xrpClient = DefaultXRPClient(networkClient: networkClient)
+    let xrpClient = DefaultXRPClient(networkClient: networkClient, xrplNetwork: XRPLNetwork.test)
 
     // WHEN the transactionHistory is requested.
     guard let transactions = try? xrpClient.paymentHistory(for: .testAddress) else {
@@ -575,7 +602,7 @@ final class DefaultXRPClientTest: XCTestCase {
       transactionStatusResult: .success(.testGetTransactionResponse),
       transactionHistoryResult: .success(transactionHistory)
     )
-    let xrpClient = DefaultXRPClient(networkClient: networkClient)
+    let xrpClient = DefaultXRPClient(networkClient: networkClient, xrplNetwork: XRPLNetwork.test)
 
     // WHEN the transactionHistory is requested THEN an error is thrown.
     XCTAssertThrowsError(try xrpClient.paymentHistory(for: .testAddress), "Exception not thrown") { error in
@@ -599,7 +626,10 @@ final class DefaultXRPClientTest: XCTestCase {
 
   func testGetPaymentWithSuccess() {
     // GIVEN a DefaultXRPClient with mocked networking that will succeed for getTransaction.
-    let xrpClient = DefaultXRPClient(networkClient: FakeNetworkClient.successfulFakeNetworkClient)
+    let xrpClient = DefaultXRPClient(
+      networkClient: FakeNetworkClient.successfulFakeNetworkClient,
+      xrplNetwork: XRPLNetwork.test
+    )
 
     // WHEN a transaction is requested.
     guard let transaction = try? xrpClient.getPayment(for: .testTransactionHash) else {
@@ -608,7 +638,13 @@ final class DefaultXRPClientTest: XCTestCase {
     }
 
     // THEN the returned transaction is as expected.
-    XCTAssertEqual(transaction, XRPTransaction(getTransactionResponse: .testGetTransactionResponse))
+    XCTAssertEqual(
+      transaction,
+      XRPTransaction(
+        getTransactionResponse: .testGetTransactionResponse,
+        xrplNetwork: XRPLNetwork.test
+      )
+    )
   }
 
   func testGetPaymentWithNotFoundError() {
@@ -630,7 +666,7 @@ final class DefaultXRPClientTest: XCTestCase {
       ),
       transactionHistoryResult: .success(.testTransactionHistoryResponse)
     )
-    let xrpClient = DefaultXRPClient(networkClient: networkClient)
+    let xrpClient = DefaultXRPClient(networkClient: networkClient, xrplNetwork: XRPLNetwork.test)
 
     // WHEN a transaction is requested, THEN the error is re-thrown.
     XCTAssertThrowsError(try xrpClient.getPayment(for: .testTransactionHash), "Exception not thrown") { error in
@@ -659,7 +695,7 @@ final class DefaultXRPClientTest: XCTestCase {
       transactionStatusResult: .success(.invalidTestGetTransactionResponse),
       transactionHistoryResult: .success(.testTransactionHistoryResponse)
     )
-    let xrpClient = DefaultXRPClient(networkClient: networkClient)
+    let xrpClient = DefaultXRPClient(networkClient: networkClient, xrplNetwork: XRPLNetwork.test)
 
     // WHEN a transaction is requested.
     let transaction = try? xrpClient.getPayment(for: .testTransactionHash)
@@ -677,7 +713,7 @@ final class DefaultXRPClientTest: XCTestCase {
       transactionStatusResult: .success(.invalidGetTransactionResponseUnsupported),
       transactionHistoryResult: .success(.testTransactionHistoryResponse)
     )
-    let xrpClient = DefaultXRPClient(networkClient: networkClient)
+    let xrpClient = DefaultXRPClient(networkClient: networkClient, xrplNetwork: XRPLNetwork.test)
 
     // WHEN a transaction is requested.
     let transaction = try? xrpClient.getPayment(for: .testTransactionHash)