+++ Tests/XRP/ProtocolBufferConversionTests.swift
@@ -296,12 +296,17 @@ final class ProtocolBufferConversionTests: XCTestCase {
     }
 
     // WHEN the protocol buffer is converted to a native Swift type.
-    let payment = XRPPayment(payment: paymentProto)
-
+    let payment = XRPPayment(payment: paymentProto, xrplNetwork: XRPLNetwork.test)
+    let expectedXAddress = Utils.encode(
+      classicAddress: payment!.destination,
+      tag: payment?.destinationTag,
+      isTest: true
+    )
     // THEN the result is as expected.
     XCTAssertEqual(payment?.amount, XRPCurrencyAmount(currencyAmount: paymentProto.amount.value))
     XCTAssertEqual(payment?.destination, paymentProto.destination.value.address)
     XCTAssertEqual(payment?.destinationTag, paymentProto.destinationTag.value)
+    XCTAssertEqual(payment?.destinationXAddress, expectedXAddress)
     XCTAssertEqual(payment?.deliverMin, XRPCurrencyAmount(currencyAmount: paymentProto.deliverMin.value))
     XCTAssertEqual(payment?.invoiceID, paymentProto.invoiceID.value)
     XCTAssertEqual(payment?.paths, paymentProto.paths.map { XRPPath(path: $0) })
@@ -324,12 +329,18 @@ final class ProtocolBufferConversionTests: XCTestCase {
     }
 
     // WHEN the protocol buffer is converted to a native Swift type.
-    let payment = XRPPayment(payment: paymentProto)
+    let payment = XRPPayment(payment: paymentProto, xrplNetwork: XRPLNetwork.test)
+    let expectedXAddress = Utils.encode(
+      classicAddress: payment!.destination,
+      tag: payment?.destinationTag,
+      isTest: true
+    )
 
     // THEN the result is as expected.
     XCTAssertEqual(payment?.amount, XRPCurrencyAmount(currencyAmount: paymentProto.amount.value))
     XCTAssertEqual(payment?.destination, paymentProto.destination.value.address)
     XCTAssertNil(payment?.destinationTag)
+    XCTAssertEqual(payment?.destinationXAddress, expectedXAddress)
     XCTAssertNil(payment?.deliverMin)
     XCTAssertNil(payment?.invoiceID)
     XCTAssertNil(payment?.paths)
@@ -352,7 +363,7 @@ final class ProtocolBufferConversionTests: XCTestCase {
     }
 
     // WHEN the protocol buffer is converted to a native Swift type THEN the result is nil
-    XCTAssertNil(XRPPayment(payment: paymentProto))
+    XCTAssertNil(XRPPayment(payment: paymentProto, xrplNetwork: XRPLNetwork.test))
   }
 
   func testConvertPaymentWithInvalidDeliverMinField() {
@@ -376,7 +387,7 @@ final class ProtocolBufferConversionTests: XCTestCase {
     }
 
     // WHEN the protocol buffer is converted to a native Swift type THEN the result is nil
-    XCTAssertNil(XRPPayment(payment: paymentProto))
+    XCTAssertNil(XRPPayment(payment: paymentProto, xrplNetwork: XRPLNetwork.test))
   }
 
   func testConvertPaymentWithInvalidSendMaxField() {
@@ -400,7 +411,7 @@ final class ProtocolBufferConversionTests: XCTestCase {
     }
 
     // WHEN the protocol buffer is converted to a native Swift type THEN the result is nil
-    XCTAssertNil(XRPPayment(payment: paymentProto))
+    XCTAssertNil(XRPPayment(payment: paymentProto, xrplNetwork: XRPLNetwork.test))
   }
 
   // MARK: - Org_Xrpl_Rpc_V1_Memo
@@ -557,7 +568,12 @@ final class ProtocolBufferConversionTests: XCTestCase {
       $0.ledgerIndex = ledgerIndex
     }
     // WHEN the protocol buffer is converted to a native Swift type.
-    let transaction = XRPTransaction(getTransactionResponse: getTransactionResponseProto)
+    let transaction = XRPTransaction(getTransactionResponse: getTransactionResponseProto, xrplNetwork: XRPLNetwork.test)
+    let expectedXAddress = Utils.encode(
+      classicAddress: transaction!.account,
+      tag: transaction!.sourceTag,
+      isTest: true
+    )
 
     // THEN all fields are present and converted correctly.
     XCTAssertEqual(transaction?.hash, [UInt8](hash).toHex())
@@ -572,6 +588,7 @@ final class ProtocolBufferConversionTests: XCTestCase {
     XCTAssertEqual(transaction?.memos, [ XRPMemo(memo: memoProto) ])
     XCTAssertEqual(transaction?.signers, [ XRPSigner(signer: signerProto) ])
     XCTAssertEqual(transaction?.sourceTag, sourceTag)
+    XCTAssertEqual(transaction?.sourceXAddress, expectedXAddress)
     XCTAssertEqual(transaction?.timestamp, .expectedTimestamp)
     XCTAssertEqual(transaction?.deliveredAmount, String(deliveredAmount))
     XCTAssertEqual(transaction?.validated, validated)
@@ -625,7 +642,16 @@ final class ProtocolBufferConversionTests: XCTestCase {
       $0.hash = hash
     }
     // WHEN the protocol buffer is converted to a native Swift type.
-    let transaction = XRPTransaction(getTransactionResponse: getTransactionResponseProto)
+    let transaction = XRPTransaction(
+      getTransactionResponse: getTransactionResponseProto,
+      xrplNetwork: XRPLNetwork.test
+    )
+
+    let expectedXAddress = Utils.encode(
+      classicAddress: transaction!.account,
+      tag: transaction!.sourceTag,
+      isTest: true
+    )
 
     // THEN all fields are present and converted correctly.
     XCTAssertEqual(transaction?.hash, [UInt8](hash).toHex())
@@ -640,6 +666,7 @@ final class ProtocolBufferConversionTests: XCTestCase {
     XCTAssertNil(transaction?.memos)
     XCTAssertNil(transaction?.signers)
     XCTAssertNil(transaction?.sourceTag)
+    XCTAssertEqual(transaction?.sourceXAddress, expectedXAddress)
     XCTAssertNil(transaction?.timestamp)
     XCTAssertNil(transaction?.deliveredAmount)
   }
@@ -680,7 +707,7 @@ final class ProtocolBufferConversionTests: XCTestCase {
       $0.hash = hash
     }
     // WHEN the protocol buffer is converted to a native Swift type.
-    let transaction = XRPTransaction(getTransactionResponse: getTransactionResponseProto)
+    let transaction = XRPTransaction(getTransactionResponse: getTransactionResponseProto, xrplNetwork: XRPLNetwork.test)
 
     // THEN the result is nil
     XCTAssertNil(transaction)
@@ -723,7 +750,7 @@ final class ProtocolBufferConversionTests: XCTestCase {
     }
 
     // WHEN the protocol buffer is converted to a native Swift type.
-    let transaction = XRPTransaction(getTransactionResponse: getTransactionResponseProto)
+    let transaction = XRPTransaction(getTransactionResponse: getTransactionResponseProto, xrplNetwork: XRPLNetwork.test)
 
     // THEN the result is nil
     XCTAssertNil(transaction)